Add more nodes implementations

+++ src/AST/integer.cpp
@@ -1,4 +1,3 @@
-
 #include "integer.h"
 
-Integer::Integer(double value) : value(value) {}
+Integer::Integer(int value) : value(value) {}
